Add test for community.py
