+++ etc/pytest.sh
@@ -12,6 +12,7 @@ mkdir -p .tmp/integration_test/lib/analysis/author/graph/generate/
 mkdir -p .tmp/integration_test/lib/analysis/author/ranking/
 mkdir -p .tmp/integration_test/lib/analysis/author/wh_table/
 mkdir -p .tmp/integration_test/lib/analysis/author/edge_list/
+mkdir -p .tmp/integration_test/lib/analysis/author/community/vertex_clustering/
 
 py.test -v --color=yes --cov=.
 codecov
